refactor: 1단계 - 문자열 계산기, 패키지 생성
- domain 패키지 생성
