+++ src/test/java/stringcalculator/domain/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package stringcalculator.domain;
 
 import static org.assertj.core.api.Assertions.*;
